refactor: Move integration test into more appropriate test module
